main/requirements.txt
streamlit==1.31.0
pymongo==4.6.1
pandas==2.2.0
matplotlib==3.8.2
python-dotenv==1.0.0
dnspython==2.5.0